refactor(SQL-IR Phase 2): retire dead Path D (LogicalExpr::to_sql)
SQL-IR Phase-2 step 3. The investigation (SQL_IR_DESIGN.md §3.5) framed this as
"collapse Path D into A via TryFrom<LogicalExpr>". A rigorous reachability audit
(independently corroborated) found there is nothing live to collapse: Path D is
ENTIRELY DEAD in production.

The decisive fact: there are two different `ToSql` traits. Production uses Path A's
(`render_plan::ToSql`, returns `String`). Path D's (`to_sql.rs::ToSql`, returns
`Result`, `#[allow(dead_code)]`) forms a closed self-referential cluster —
`to_sql.rs` <-> `function_translator::translate_scalar_function` <-> `view_query.rs`
— whose only doors in are its own recursion, `#[cfg(test)]` blocks, and
`build_view_scan` (zero callers). No production entry point (server, embedded, FFI,
CLI) reaches it: Path A's `RenderExpr::ScalarFnCall` arm has its own self-contained
renderer and never calls `translate_scalar_function`.

Deleted (all dead):
- `to_sql.rs` (whole file: the `ToSql` trait, its LogicalExpr/LogicalPlan/Arc impls,
  6 `_logical` helpers, its test module) — 543 lines
- `view_query.rs` (dead `PlanViewScan` renderer) + `view_scan.rs` (`build_view_scan`,
  zero callers)
- `function_translator::translate_scalar_function` + `translate_duration_function`
  + their 16 unit tests + the `translate_scalar_function` re-export in mod.rs.
  KEPT the live members: `interval_expr_for_unit` (used by Path A), the `ch.`/`chagg.`
  consts + `is_ch_aggregate_function` + `CH_AGGREGATE_FUNCTIONS`, and their tests.

Coverage preserved (no net loss):
- Ported the dead `to_sql.rs` FINAL dialect-gating test to the LIVE path
  (`FromTableItem::to_sql`, `to_sql_query.rs` tests) — same CH-emits/DBX-omits
  contract via `supports_final_keyword()`.
- The deleted `ch.`/`chagg.` passthrough tests are redundant: `strip_passthrough`
  has 12 dedicated tests in `sql_generator/passthrough/`. Duration + function-map
  transforms are covered by `interval_expr_for_unit_*` unit tests + `interval_*`
  goldens + `function_registry` tests.
- Closed the ONE genuine gap the cleanup surfaced: added an `fn_datetime_epoch_millis`
  sql_ir golden locking `datetime({epochMillis: x})` -> identity pass-through (the
  removed Path-D test was its only prior coverage). Renders `x` as-is in both dialects.
- Refreshed two now-stale comments that cited `translate_scalar_function` /
  "the LogicalExpr arms" as the message-bearing error path.

Net -1112 lines. ClickHouse + Databricks byte-identical (only the 2 new datetime
goldens added, no churn). Full gate: fmt + clippy (no dead_code) + 1596 lib +
sql_golden + corpus_sweep + ratchet net-zero, in an isolated CARGO_TARGET_DIR
(shared target lock-contended by another project).
