--- src/sql_generator/emitters/clickhouse/to_sql.rs
@@ -1,543 +1,0 @@
-use super::errors::ClickhouseQueryGeneratorError;
-use super::function_registry::get_function_mapping;
-use super::function_translator::translate_scalar_function;
-use crate::query_planner::logical_expr::{Literal, LogicalExpr, Operator};
-use crate::query_planner::logical_plan::LogicalPlan;
-use std::sync::Arc;
-
-/// Check if an expression contains a string literal (recursively for nested + operations)
-fn contains_string_literal_logical(expr: &LogicalExpr) -> bool {
-    match expr {
-        LogicalExpr::Literal(Literal::String(_)) => true,
-        LogicalExpr::OperatorApplicationExp(op) if op.operator == Operator::Addition => {
-            op.operands.iter().any(contains_string_literal_logical)
-        }
-        _ => false,
-    }
-}
-
-/// Check if any operand in the expression is a string literal
-fn has_string_operand_logical(operands: &[LogicalExpr]) -> bool {
-    operands.iter().any(contains_string_literal_logical)
-}
-
-/// Check if a LogicalExpr is a list/array expression (for arrayConcat detection).
-fn is_list_expr_logical(expr: &LogicalExpr) -> bool {
-    match expr {
-        LogicalExpr::AggregateFnCall(agg) => {
-            agg.name.eq_ignore_ascii_case("collect") || agg.name.eq_ignore_ascii_case("groupArray")
-        }
-        LogicalExpr::ScalarFnCall(f) => {
-            let n = f.name.to_lowercase();
-            n == "arrayconcat" || n == "arraysort" || n == "arraydistinct"
-        }
-        LogicalExpr::List(_) => true,
-        LogicalExpr::OperatorApplicationExp(op) if op.operator == Operator::Addition => {
-            op.operands.iter().any(is_list_expr_logical)
-        }
-        _ => false,
-    }
-}
-
-/// Flatten nested + operations into a list of SQL strings for arrayConcat().
-/// Known scalar literals are wrapped as `[scalar]` so that `list + scalar`
-/// produces valid `arrayConcat(list, [scalar])` (ClickHouse requires array args).
-/// Ambiguous expressions (Identifier, PropertyAccessExp, etc.) are left as-is
-/// since they may already hold arrays.
-fn flatten_list_addition_operands_logical(
-    expr: &LogicalExpr,
-) -> Result<Vec<String>, ClickhouseQueryGeneratorError> {
-    match expr {
-        LogicalExpr::OperatorApplicationExp(op) if op.operator == Operator::Addition => {
-            let mut result = Vec::new();
-            for operand in &op.operands {
-                result.extend(flatten_list_addition_operands_logical(operand)?);
-            }
-            Ok(result)
-        }
-        _ => {
-            let sql = ToSql::to_sql(expr)?;
-            if is_known_scalar_logical(expr) {
-                // Wrap scalar as single-element array for arrayConcat/concat compatibility
-                Ok(vec![
-                    crate::sql_generator::function_mapper::current_function_mapper()
-                        .array_literal(&sql),
-                ])
-            } else {
-                Ok(vec![sql])
-            }
-        }
-    }
-}
-
-/// Returns true if the expression is definitely a scalar (not an array).
-fn is_known_scalar_logical(expr: &LogicalExpr) -> bool {
-    matches!(
-        expr,
-        LogicalExpr::Literal(Literal::Integer(_))
-            | LogicalExpr::Literal(Literal::Float(_))
-            | LogicalExpr::Literal(Literal::String(_))
-            | LogicalExpr::Literal(Literal::Boolean(_))
-            | LogicalExpr::Literal(Literal::Null)
-            | LogicalExpr::Parameter(_)
-    )
-}
-
-/// Flatten nested + operations into a list of SQL strings for concat()
-fn flatten_addition_operands_logical(
-    expr: &LogicalExpr,
-) -> Result<Vec<String>, ClickhouseQueryGeneratorError> {
-    match expr {
-        LogicalExpr::OperatorApplicationExp(op) if op.operator == Operator::Addition => {
-            let mut result = Vec::new();
-            for operand in &op.operands {
-                result.extend(flatten_addition_operands_logical(operand)?);
-            }
-            Ok(result)
-        }
-        _ => {
-            // Use the ToSql trait impl for LogicalExpr
-            Ok(vec![ToSql::to_sql(expr)?])
-        }
-    }
-}
-
-/// Convert a plan node to SQL
-#[allow(dead_code)]
-pub trait ToSql {
-    /// Convert this node to a SQL string
-    fn to_sql(&self) -> Result<String, ClickhouseQueryGeneratorError>;
-}
-
-impl ToSql for Arc<LogicalPlan> {
-    fn to_sql(&self) -> Result<String, ClickhouseQueryGeneratorError> {
-        self.as_ref().to_sql()
-    }
-}
-
-impl ToSql for LogicalExpr {
-    fn to_sql(&self) -> Result<String, ClickhouseQueryGeneratorError> {
-        match self {
-            LogicalExpr::Literal(lit) => match lit {
-                Literal::Integer(i) => Ok(i.to_string()),
-                Literal::Float(f) => Ok(f.to_string()),
-                Literal::Boolean(b) => Ok(b.to_string()),
-                Literal::String(s) => Ok(format!("'{}'", s)),
-                Literal::Null => Ok("NULL".to_string()),
-            },
-            LogicalExpr::Raw(raw) => Ok(raw.clone()),
-            LogicalExpr::Star => Ok("*".into()),
-            LogicalExpr::TableAlias(alias) => Ok(alias.0.clone()),
-            LogicalExpr::ColumnAlias(col) => Ok(col.0.clone()),
-            LogicalExpr::Column(col) => Ok(col.0.clone()),
-            LogicalExpr::Parameter(param) => Ok(format!("${}", param)),
-            LogicalExpr::List(items) => {
-                // Use tuple/struct for comparison to handle mixed types
-                // (date, int, string, etc.). CH=tuple, Spark=struct.
-                let items_sql: Result<Vec<String>, _> = items.iter().map(|e| e.to_sql()).collect();
-                let ctor = crate::sql_generator::function_mapper::current_function_mapper()
-                    .tuple_constructor();
-                Ok(format!("{}({})", ctor, items_sql?.join(", ")))
-            }
-            LogicalExpr::AggregateFnCall(fn_call) => {
-                let args_sql: Result<Vec<String>, _> =
-                    fn_call.args.iter().map(|e| e.to_sql()).collect();
-                let args_sql = args_sql?;
-
-                // Native-function pass-through (ch./chagg. for ClickHouse, dbx. for
-                // Databricks). This arm returns Result, so a foreign-backend prefix
-                // is surfaced as a clean translation error here.
-                if let Some(bare) = crate::sql_generator::passthrough::strip_passthrough(
-                    &fn_call.name,
-                    crate::server::query_context::get_current_dialect(),
-                )
-                .map_err(|e| ClickhouseQueryGeneratorError::SchemaError(e.to_string()))?
-                {
-                    return Ok(format!("{}({})", bare, args_sql.join(", ")));
-                }
-
-                // Use function registry to translate Neo4j -> ClickHouse function names
-                // Note: We use ClickHouse-specific aggregate names (like "anyLast") to avoid
-                // conflicts with Cypher functions (like "any" for array predicates)
-                let fn_name_lower = fn_call.name.to_lowercase();
-
-                // percentileCont/Disc are parametric quantiles — render through the
-                // dialect FunctionMapper, honoring the percentile arg (#639).
-                if let Some(sql) = super::common::try_render_percentile(&fn_name_lower, &args_sql) {
-                    return Ok(sql);
-                }
-
-                if let Some(mapping) = get_function_mapping(&fn_name_lower) {
-                    // Apply argument transformation if provided
-                    let transformed_args = if let Some(transform_fn) = mapping.arg_transform {
-                        transform_fn(&args_sql)
-                    } else {
-                        args_sql
-                    };
-                    Ok(format!(
-                        "{}({})",
-                        mapping.name_for(crate::server::query_context::get_current_dialect()),
-                        transformed_args.join(", ")
-                    ))
-                } else {
-                    // No mapping, use name directly (standard SQL functions like count, sum, etc.)
-                    Ok(format!("{}({})", fn_call.name, args_sql.join(", ")))
-                }
-            }
-            LogicalExpr::ScalarFnCall(fn_call) => {
-                // Use function translator for Neo4j -> ClickHouse mapping
-                translate_scalar_function(fn_call)
-            }
-            LogicalExpr::PropertyAccessExp(prop) => {
-                // Property has been resolved from schema during query planning.
-                // PropertyValue already contains the correct mapping:
-                // - Column(_): Direct column mapping (e.g., year: Year)
-                // - Expression(_): Expression mapping (e.g., year: toYear(FlightDate))
-                // Just render it as-is.
-                Ok(prop.column.to_sql(&prop.table_alias.0))
-            }
-            LogicalExpr::OperatorApplicationExp(op) => {
-                // ⚠️ TODO: Operator rendering consolidation (SQL-IR Phase 2 step 3).
-                // This block duplicates the operator handling in to_sql_query.rs.
-                // The former blocker — two distinct `Operator` types — is GONE: the
-                // enums were unified (render_expr::Operator now re-exports
-                // logical_expr::Operator, PR #818). What remains: this block walks
-                // `LogicalExpr` operands and returns `Result`, while the render-side
-                // block walks `RenderExpr` operands and returns `String`. Collapsing
-                // is now the Path D -> A retirement (SQL_IR_DESIGN.md §3.5 step 3):
-                // bake LogicalExpr -> RenderExpr at this block's callers (the existing
-                // `TryFrom<LogicalExpr>`), then render via the one RenderExpr path.
-                let operands_sql: Vec<String> = op
-                    .operands
-                    .iter()
-                    .map(|e| e.to_sql())
-                    .collect::<Result<Vec<String>, _>>()?;
-                match op.operator {
-                    Operator::Addition => {
-                        // Use arrayConcat() for list concatenation
-                        if op.operands.iter().any(is_list_expr_logical) {
-                            let flattened: Vec<String> = op
-                                .operands
-                                .iter()
-                                .map(flatten_list_addition_operands_logical)
-                                .collect::<Result<Vec<Vec<String>>, _>>()?
-                                .into_iter()
-                                .flatten()
-                                .collect();
-                            Ok(format!(
-                                "{}({})",
-                                crate::sql_generator::function_mapper::current_function_mapper()
-                                    .array_concat(),
-                                flattened.join(", ")
-                            ))
-                        }
-                        // Use concat() for string concatenation, + for numeric
-                        // Flatten nested + operations for cases like: a + ' - ' + b
-                        else if has_string_operand_logical(&op.operands) {
-                            let flattened: Vec<String> = op
-                                .operands
-                                .iter()
-                                .map(flatten_addition_operands_logical)
-                                .collect::<Result<Vec<Vec<String>>, _>>()?
-                                .into_iter()
-                                .flatten()
-                                .collect();
-                            Ok(format!("concat({})", flattened.join(", ")))
-                        } else {
-                            Ok(format!("({} + {})", operands_sql[0], operands_sql[1]))
-                        }
-                    }
-                    Operator::Subtraction => {
-                        Ok(format!("({} - {})", operands_sql[0], operands_sql[1]))
-                    }
-                    Operator::Multiplication => {
-                        Ok(format!("({} * {})", operands_sql[0], operands_sql[1]))
-                    }
-                    Operator::Division => {
-                        Ok(format!("({} / {})", operands_sql[0], operands_sql[1]))
-                    }
-                    Operator::ModuloDivision => {
-                        Ok(format!("({} % {})", operands_sql[0], operands_sql[1]))
-                    }
-                    Operator::Exponentiation => {
-                        Ok(format!("power({}, {})", operands_sql[0], operands_sql[1]))
-                    }
-                    Operator::Equal => Ok(format!("({} = {})", operands_sql[0], operands_sql[1])),
-                    Operator::NotEqual => {
-                        Ok(format!("({} != {})", operands_sql[0], operands_sql[1]))
-                    }
-                    Operator::LessThan => {
-                        Ok(format!("({} < {})", operands_sql[0], operands_sql[1]))
-                    }
-                    Operator::GreaterThan => {
-                        Ok(format!("({} > {})", operands_sql[0], operands_sql[1]))
-                    }
-                    Operator::LessThanEqual => {
-                        Ok(format!("({} <= {})", operands_sql[0], operands_sql[1]))
-                    }
-                    Operator::GreaterThanEqual => {
-                        Ok(format!("({} >= {})", operands_sql[0], operands_sql[1]))
-                    }
-                    Operator::RegexMatch => {
-                        // CH match() / Spark rlike() — dialect-aware.
-                        Ok(super::common::regex_match_predicate(
-                            &operands_sql[0],
-                            &operands_sql[1],
-                        ))
-                    }
-                    Operator::And => Ok(format!("({} AND {})", operands_sql[0], operands_sql[1])),
-                    Operator::Or => Ok(format!("({} OR {})", operands_sql[0], operands_sql[1])),
-                    Operator::In => {
-                        // Cypher: x IN array_property → CH: has(arr, x), Spark: array_contains(arr, x)
-                        // Cypher: x IN [1, 2, 3]    → standard IN list
-                        if matches!(&op.operands[1], LogicalExpr::PropertyAccessExp(_)) {
-                            let contains =
-                                crate::sql_generator::function_mapper::current_function_mapper()
-                                    .array_contains();
-                            Ok(format!(
-                                "{}({}, {})",
-                                contains, operands_sql[1], operands_sql[0]
-                            ))
-                        } else {
-                            Ok(format!("({} IN {})", operands_sql[0], operands_sql[1]))
-                        }
-                    }
-                    Operator::NotIn => {
-                        if matches!(&op.operands[1], LogicalExpr::PropertyAccessExp(_)) {
-                            let contains =
-                                crate::sql_generator::function_mapper::current_function_mapper()
-                                    .array_contains();
-                            Ok(format!(
-                                "NOT {}({}, {})",
-                                contains, operands_sql[1], operands_sql[0]
-                            ))
-                        } else {
-                            // Literal list: standard NOT IN
-                            Ok(format!("({} NOT IN {})", operands_sql[0], operands_sql[1]))
-                        }
-                    }
-                    Operator::StartsWith => {
-                        // Dialect-aware: CH startsWith vs Spark startswith.
-                        Ok(super::common::starts_with_predicate(
-                            &operands_sql[0],
-                            &operands_sql[1],
-                        ))
-                    }
-                    Operator::EndsWith => {
-                        // Dialect-aware: CH endsWith vs Spark endswith.
-                        Ok(super::common::ends_with_predicate(
-                            &operands_sql[0],
-                            &operands_sql[1],
-                        ))
-                    }
-                    Operator::Contains => {
-                        // position(haystack, needle) > 0 — dialect-aware (Spark reverses args).
-                        Ok(super::common::contains_predicate(
-                            &operands_sql[0],
-                            &operands_sql[1],
-                        ))
-                    }
-                    Operator::Not => Ok(format!("NOT ({})", operands_sql[0])),
-                    Operator::Distinct => Ok(format!("DISTINCT {}", operands_sql[0])),
-                    Operator::IsNull => Ok(format!("({} IS NULL)", operands_sql[0])),
-                    Operator::IsNotNull => Ok(format!("({} IS NOT NULL)", operands_sql[0])),
-                }
-            }
-            LogicalExpr::Case(case_expr) => {
-                let mut sql = String::from("CASE ");
-
-                // Simple CASE (CASE x WHEN ...)
-                if let Some(expr) = &case_expr.expr {
-                    sql.push_str(&format!("{} ", expr.to_sql()?));
-                }
-
-                // WHEN ... THEN ... clauses
-                for (when_expr, then_expr) in &case_expr.when_then {
-                    sql.push_str(&format!(
-                        "WHEN {} THEN {} ",
-                        when_expr.to_sql()?,
-                        then_expr.to_sql()?
-                    ));
-                }
-
-                // ELSE clause
-                if let Some(else_expr) = &case_expr.else_expr {
-                    sql.push_str(&format!("ELSE {} ", else_expr.to_sql()?));
-                }
-
-                sql.push_str("END");
-                Ok(sql)
-            }
-            LogicalExpr::InSubquery(in_subquery) => {
-                let expr_sql = in_subquery.expr.to_sql()?;
-                let subquery_sql = in_subquery.subplan.to_sql()?;
-                Ok(format!("{} IN ({})", expr_sql, subquery_sql))
-            }
-            LogicalExpr::ExistsSubquery(exists_subquery) => {
-                // Generate EXISTS (SELECT 1 FROM ... WHERE ...)
-                let subquery_sql = exists_subquery.subplan.to_sql()?;
-                Ok(format!("EXISTS ({})", subquery_sql))
-            }
-            LogicalExpr::ReduceExpr(reduce) => {
-                // Convert Cypher reduce(acc = init, x IN list | expr) to ClickHouse arrayFold
-                // ClickHouse syntax: arrayFold((acc, x) -> expr, list, init)
-                // Note: Cast init to Int64 to avoid type mismatch issues when the lambda
-                // expression returns a wider type than the inferred initial value type
-                let init_sql = reduce.initial_value.to_sql()?;
-                let list_sql = reduce.list.to_sql()?;
-                let expr_sql = reduce.expression.to_sql()?;
-
-                // Wrap numeric init values in a 64-bit-int cast to prevent
-                // type mismatch when the lambda returns a wider type.
-                let init_cast = if matches!(
-                    *reduce.initial_value,
-                    LogicalExpr::Literal(Literal::Integer(_))
-                ) {
-                    let fmap = crate::sql_generator::function_mapper::current_function_mapper();
-                    format!("{}({})", fmap.cast_int64(), init_sql)
-                } else {
-                    init_sql
-                };
-
-                Ok(super::common::reduce_fold_sql(
-                    &reduce.variable,
-                    &reduce.accumulator,
-                    &expr_sql,
-                    &list_sql,
-                    &init_cast,
-                ))
-            }
-            LogicalExpr::MapLiteral(entries) => {
-                // Use ClickHouse map() function for map literals
-                // map('key1', val1, 'key2', val2, ...)
-                //
-                // IMPORTANT: ClickHouse requires all map values to be of the same type.
-                // Since Cypher maps can have mixed types (e.g., {name:'nodes', data:count(*)}),
-                // we cast all values to String to ensure type compatibility.
-                if entries.is_empty() {
-                    Ok("map()".to_string())
-                } else {
-                    let to_str = crate::sql_generator::function_mapper::current_function_mapper()
-                        .cast_string();
-                    let args: Result<Vec<String>, _> = entries
-                        .iter()
-                        .flat_map(|(k, v)| {
-                            let val_result = v.to_sql();
-                            vec![
-                                Ok(format!("'{}'", k)),
-                                val_result.map(|val| format!("{}({})", to_str, val)),
-                            ]
-                        })
-                        .collect();
-                    Ok(format!("map({})", args?.join(", ")))
-                }
-            }
-            LogicalExpr::LabelExpression { variable, label } => {
-                // Label expression should typically be resolved at planning time
-                // If we reach here, it means we couldn't determine the label statically
-                // Generate a fallback that returns false (unknown label)
-                // In a more sophisticated implementation, this could query a type column
-                log::warn!(
-                    "LabelExpression {}:{} reached SQL generation - returning false",
-                    variable,
-                    label
-                );
-                Ok("false".to_string())
-            }
-            LogicalExpr::PathPattern(_) => {
-                // Path patterns are handled at the logical plan level, not expression level
-                Err(ClickhouseQueryGeneratorError::UnsupportedItemInWhereClause)
-            }
-            _ => todo!("Not yet implemented"),
-        }
-    }
-}
-
-impl ToSql for LogicalPlan {
-    fn to_sql(&self) -> Result<String, ClickhouseQueryGeneratorError> {
-        match self {
-            LogicalPlan::ViewScan(scan) => {
-                // Add FINAL keyword if enabled. FINAL is ClickHouse-only — never emit it
-                // on other dialects (e.g. Databricks/Spark) where it is invalid SQL.
-                let final_keyword = if scan.use_final
-                    && crate::server::query_context::get_current_dialect().supports_final_keyword()
-                {
-                    " FINAL"
-                } else {
-                    ""
-                };
-                let mut sql = format!("SELECT * FROM {}{}", scan.source_table, final_keyword);
-
-                // Add WHERE clause if view_filter is present
-                if let Some(ref filter) = scan.view_filter {
-                    sql.push_str(" WHERE ");
-                    sql.push_str(&filter.to_sql()?);
-                }
-                Ok(sql)
-            }
-            LogicalPlan::Projection(proj) => {
-                let mut sql = String::new();
-                sql.push_str("SELECT ");
-
-                let mut projections = Vec::new();
-                for item in &proj.items {
-                    projections.push(item.expression.to_sql()?);
-                }
-                sql.push_str(&projections.join(", "));
-
-                sql.push_str("\nFROM ");
-                sql.push_str(&proj.input.to_sql()?);
-
-                Ok(sql)
-            }
-            _ => Err(ClickhouseQueryGeneratorError::UnsupportedDDLQuery),
-        }
-    }
-}
-
-#[cfg(test)]
-mod final_dialect_tests {
-    use super::ToSql;
-    use crate::query_planner::logical_plan::{LogicalPlan, ViewScan};
-    use crate::server::query_context::{with_query_context, QueryContext};
-    use crate::sql_generator::SqlDialect;
-    use std::collections::HashMap;
-    use std::sync::Arc;
-
-    fn view_scan_with_final() -> LogicalPlan {
-        let mut scan = ViewScan::new(
-            "users".to_string(),
-            None,
-            HashMap::new(),
-            "id".to_string(),
-            vec![],
-            vec![],
-        );
-        scan.use_final = true;
-        LogicalPlan::ViewScan(Arc::new(scan))
-    }
-
-    /// FINAL is ClickHouse-only: emitted under ClickHouse, omitted under Databricks
-    /// even when the schema requests it (use_final = true).
-    #[tokio::test]
-    async fn final_emitted_only_on_clickhouse() {
-        // Default scope -> ClickHouse -> FINAL present.
-        let ch_sql = view_scan_with_final().to_sql().unwrap();
-        assert!(
-            ch_sql.contains(" FINAL"),
-            "ClickHouse should emit FINAL; got: {ch_sql}"
-        );
-
-        // Databricks dialect -> no FINAL (invalid Spark SQL).
-        let ctx = QueryContext {
-            dialect: SqlDialect::Databricks,
-            ..QueryContext::default()
-        };
-        let dbx_sql =
-            with_query_context(ctx, async { view_scan_with_final().to_sql().unwrap() }).await;
-        assert!(
-            !dbx_sql.contains("FINAL"),
-            "Databricks must not emit FINAL; got: {dbx_sql}"
-        );
-    }
-}
