--- src/sql_generator/emitters/clickhouse/view_scan.rs
@@ -1,58 +1,0 @@
-//! ViewScan operations support
-//!
-//! This module handles converting ViewScan operations from the logical plan
-//! into equivalent SQL queries.
-
-use crate::query_planner::logical_plan::{LogicalPlan, ViewScan};
-
-/// Build a SQL query for a ViewScan operation
-/// Reserved for future use when standalone ViewScan SQL generation is needed
-#[allow(dead_code)]
-pub fn build_view_scan(scan: &ViewScan, _plan: &LogicalPlan) -> String {
-    let mut sql = String::new();
-
-    // Build table reference with parameters if this is a parameterized view
-    let table_ref = if let (Some(param_names), Some(param_values)) =
-        (&scan.view_parameter_names, &scan.view_parameter_values)
-    {
-        // This is a parameterized view - generate view(param=value, ...) syntax
-        let param_pairs: Vec<String> = param_names
-            .iter()
-            .filter_map(|name| {
-                param_values.get(name).map(|value| {
-                    // Escape single quotes in value for SQL safety
-                    let escaped_value = value.replace('\'', "''");
-                    format!("{} = '{}'", name, escaped_value)
-                })
-            })
-            .collect();
-
-        if param_pairs.is_empty() {
-            // No matching parameters found - use plain table/view name
-            log::warn!(
-                "ViewScan: View '{}' expects parameters {:?} but none matched in provided values",
-                scan.source_table,
-                param_names
-            );
-            scan.source_table.clone()
-        } else {
-            // Generate parameterized view call: view_name(param1='value1', param2='value2')
-            format!("{}({})", scan.source_table, param_pairs.join(", "))
-        }
-    } else {
-        // Not a parameterized view - use plain table/view name
-        scan.source_table.clone()
-    };
-
-    // Add FINAL keyword if enabled for this table. FINAL is ClickHouse-only —
-    // never emit it on other dialects (e.g. Databricks/Spark) where it is invalid SQL.
-    if scan.use_final
-        && crate::server::query_context::get_current_dialect().supports_final_keyword()
-    {
-        sql.push_str(&format!("SELECT * FROM {} FINAL", table_ref));
-    } else {
-        sql.push_str(&format!("SELECT * FROM {}", table_ref));
-    }
-
-    sql
-}
