+++ tests/rust/integration/golden/sql_ir/standard/fn_datetime_epoch_millis__databricks.sql
@@ -1,0 +1,3 @@
+SELECT 
+      u.registration_date AS `d`
+FROM social.users_bench AS u
